+++ src/FunFair.BuildCheck/IProjectLoader.cs
@@ -1,0 +1,17 @@
+﻿using System.Xml;
+
+namespace FunFair.BuildCheck
+{
+    /// <summary>
+    ///     Loads projects.
+    /// </summary>
+    public interface IProjectLoader
+    {
+        /// <summary>
+        ///     Loads the named project.
+        /// </summary>
+        /// <param name="path">The project.</param>
+        /// <returns>The loaded project</returns>
+        XmlDocument Load(string path);
+    }
+}