+++ src/FunFair.BuildCheck/ProjectChecks/CodeAnalysisTreatWarningsAsErrorsPolicy.cs
@@ -26,13 +26,9 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
-            ProjectValueHelpers.CheckValue(projectName: projectName,
-                                           project: project,
-                                           nodePresence: @"CodeAnalysisTreatWarningsAsErrors",
-                                           requiredValue: EXPECTED,
-                                           logger: this._logger);
+            ProjectValueHelpers.CheckValue(projectName: projectName, project: project, nodePresence: @"CodeAnalysisTreatWarningsAsErrors", requiredValue: EXPECTED, logger: this._logger);
         }
     }
 }