+++ src/FunFair.BuildCheck/ProjectChecks/DoesNotUseDotNetCliToolReference.cs
@@ -23,7 +23,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             XmlNodeList nodes = project.SelectNodes(xpath: "/Project/ItemGroup/DotNetCliToolReference");