+++ src/FunFair.BuildCheck/ProjectChecks/EnableNetAnalyzersPolicy.cs
@@ -26,7 +26,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             ProjectValueHelpers.CheckValue(projectName: projectName, project: project, nodePresence: @"EnableNETAnalyzers", requiredValue: EXPECTED, logger: this._logger);
         }