+++ src/FunFair.BuildCheck/ProjectChecks/HasConsistentNuGetPackages.cs
@@ -29,7 +29,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             XmlNodeList nodes = project.SelectNodes(xpath: "/Project/ItemGroup/PackageReference");