+++ src/FunFair.BuildCheck/ProjectChecks/MustHaveFxCopAnalyzerPackage.cs
@@ -29,7 +29,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             if (this.CheckPackageReference(projectName: projectName, project: project))
             {