+++ src/FunFair.BuildCheck/ProjectChecks/MustNotDisableUnexpectedWarnings.cs
@@ -34,7 +34,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             bool isTestProject = project.IsTestProject(projectName: projectName, logger: this._logger);