+++ src/FunFair.BuildCheck/ProjectChecks/MustNotReferencePackages.cs
@@ -28,7 +28,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             foreach (string packageId in this._packageIds)
             {