+++ src/FunFair.BuildCheck/ProjectChecks/MustSpecifyOutputType.cs
@@ -24,7 +24,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             const string msg = "Exe or Library";
 
@@ -38,12 +38,7 @@ namespace FunFair.BuildCheck.ProjectChecks
                 return StringComparer.InvariantCultureIgnoreCase.Equals(x: value, y: "Exe") || StringComparer.InvariantCultureIgnoreCase.Equals(x: value, y: "Library");
             }
 
-            ProjectValueHelpers.CheckValue(projectName: projectName,
-                                           project: project,
-                                           nodePresence: @"OutputType",
-                                           isRequiredValue: IsRequiredValue,
-                                           msg: msg,
-                                           logger: this._logger);
+            ProjectValueHelpers.CheckValue(projectName: projectName, project: project, nodePresence: @"OutputType", isRequiredValue: IsRequiredValue, msg: msg, logger: this._logger);
         }
     }
 }