+++ src/FunFair.BuildCheck/ProjectChecks/NuGetPolicyDisableImplicitNuGetFallbackFolder.cs
@@ -24,13 +24,9 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
-            ProjectValueHelpers.CheckValue(projectName: projectName,
-                                           project: project,
-                                           nodePresence: @"DisableImplicitNuGetFallbackFolder",
-                                           requiredValue: "true",
-                                           logger: this._logger);
+            ProjectValueHelpers.CheckValue(projectName: projectName, project: project, nodePresence: @"DisableImplicitNuGetFallbackFolder", requiredValue: "true", logger: this._logger);
         }
     }
 }