+++ src/FunFair.BuildCheck/ProjectChecks/ReferencesNugetPackageOnlyOnce.cs
@@ -25,7 +25,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         }
 
         /// <inheritdoc />
-        public void Check(string projectName, XmlDocument project)
+        public void Check(string projectName, string projectFolder, XmlDocument project)
         {
             HashSet<string> packageReferences = new HashSet<string>(StringComparer.OrdinalIgnoreCase);